Move default_celery.py to inside the provider
This has been missed in #32526 and is extracted out from #32604
in an attempt to make it smaller and separately reviewable.

This one adds also deprecation warning to handle the
configuration value that people might already have in
the [celery] iccelery_config_options"
