+++ airflow/config_templates/default_airflow.cfg
@@ -1109,7 +1109,7 @@ flower_basic_auth =
 sync_parallelism = 0
 
 # Import path for celery configuration options
-celery_config_options = airflow.config_templates.default_celery.DEFAULT_CELERY_CONFIG
+celery_config_options = airflow.providers.celery.executors.default_celery.DEFAULT_CELERY_CONFIG
 ssl_active = False
 
 # Path to the client key.
