+++ airflow/providers/celery/executors/default_celery.py
@@ -74,11 +74,16 @@ DEFAULT_CELERY_CONFIG = {
     "worker_enable_remote_control": conf.getboolean("celery", "worker_enable_remote_control"),
 }
 
-celery_ssl_active = False
-try:
-    celery_ssl_active = conf.getboolean("celery", "SSL_ACTIVE")
-except AirflowConfigException:
-    log.warning("Celery Executor will run without SSL")
+
+def _get_celery_ssl_active() -> bool:
+    try:
+        return conf.getboolean("celery", "SSL_ACTIVE")
+    except AirflowConfigException:
+        log.warning("Celery Executor will run without SSL")
+        return False
+
+
+celery_ssl_active = _get_celery_ssl_active()
 
 try:
     if celery_ssl_active:
